传递栏目id修改为Long型

+++ detailproject/src/main/java/com/zjrb/zjxw/detailproject/link/LiveLinkActivity.java
@@ -581,7 +581,7 @@ public class LiveLinkActivity extends BaseActivity implements View.OnClickListen
     private void SyncSubscribeColumn(boolean isSubscribe, int columnid) {
         Intent intent = new Intent("subscribe_success");
         intent.putExtra("subscribe", isSubscribe);
-        intent.putExtra("id", columnid);
+        intent.putExtra("id", (long) columnid);
         LocalBroadcastManager.getInstance(this).sendBroadcast(intent);
     }
