+++ detailproject/src/main/java/com/zjrb/zjxw/detailproject/redBoat/RedBoatActivity.java
@@ -377,7 +377,7 @@ public class RedBoatActivity extends BaseActivity implements View.OnClickListene
     private void SyncSubscribeColumn(boolean isSubscribe, int columnid) {
         Intent intent = new Intent("subscribe_success");
         intent.putExtra("subscribe", isSubscribe);
-        intent.putExtra("id", columnid);
+        intent.putExtra("id", (long) columnid);
         LocalBroadcastManager.getInstance(this).sendBroadcast(intent);
     }
