+++ detailproject/src/main/java/com/zjrb/zjxw/detailproject/topic/ActivityTopicActivity.java
@@ -672,7 +672,7 @@ public class ActivityTopicActivity extends BaseActivity implements
     private void SyncSubscribeColumn(boolean isSubscribe, int columnid) {
         Intent intent = new Intent("subscribe_success");
         intent.putExtra("subscribe", isSubscribe);
-        intent.putExtra("id", columnid);
+        intent.putExtra("id", (long )columnid);
         LocalBroadcastManager.getInstance(this).sendBroadcast(intent);
     }
